Old Student dashboard files and their CSS deleted | removed colleges.js & Card.js
