--- src/components/colleges.js
@@ -1,919 +1,0 @@
-const colleges = {
-  0: "Abhilashi University",
-  1: "Acharya N.G. Ranga Agricultural University",
-  2: "Acharya Nagarjuna University",
-  3: "Adamas University",
-  4: "Adesh University",
-  5: "Adikavi Nannaya University",
-  6: "Ahmedabad University",
-  7: "AISECT University",
-  8: "Ajeenkya D.Y. Patil University",
-  9: "Akal University",
-  10: "AKS University",
-  11: "Alagappa University",
-  12: "Al-Falah University",
-  13: "Aliah University",
-  14: "Aligarh Muslim University",
-  15: "All India Institute of Medical Sciences Bhopal",
-  16: "All India Institute of Medical Sciences Bhubaneswar",
-  17: "All India Institute of Medical Sciences Delhi",
-  18: "All India Institute of Medical Sciences Jodhpur",
-  19: "All India Institute of Medical Sciences Patna",
-  20: "All India Institute of Medical Sciences Raipur",
-  21: "All India Institute of Medical Sciences Rishikesh",
-  22: "Allahabad State University",
-  23: "Alliance University",
-  24: "Ambedkar University Delhi",
-  25: "Amity University",
-  26: "Amrita Vishwa Vidyapeetham",
-  27: "Anand Agricultural University",
-  28: "Anant National University",
-  29: "Andhra University",
-  30: "Anna University",
-  31: "Annamalai University",
-  32: "Ansal University",
-  33: "AP Goyal Shimla University",
-  34: "Apeejay Stya University",
-  35: "Apex Professional University",
-  36: "APJ Abdul Kalam Technological University",
-  37: "Arka Jain University",
-  38: "Arni University",
-  39: "Arunachal University of Studies",
-  40: "Arunodaya University",
-  41: "Aryabhatta Knowledge University",
-  42: "Ashoka University",
-  43: "Assam Agricultural University",
-  44: "Assam Don Bosco University",
-  45: "Assam Down Town University",
-  46: "Assam Rajiv Gandhi University of Cooperative Management",
-  47: "Assam Science and Technology University",
-  48: "Assam University",
-  49: "Assam Women's University",
-  50: "Atal Bihari Vajpayee Hindi Vishwavidyalaya",
-  51: "AURO University",
-  52: "Avantika University",
-  53: "Avinashilingam University",
-  54: "Awadhesh Pratap Singh University",
-  55: "Ayush and Health Sciences University of Chhattisgarh",
-  56: "Azim Premji University",
-  57: "B.S. Abdur Rahman University",
-  58: "Baba Farid University of Health Sciences",
-  59: "Baba Ghulam Shah Badhshah University",
-  60: "Baba Mastnath University",
-  61: "Babasaheb Bhimrao Ambedkar Bihar University",
-  62: "Babasaheb Bhimrao Ambedkar University",
-  63: "Babu Banarasi Das University",
-  64: "Baddi University of Emerging Sciences and Technologies",
-  65: "BAHRA University",
-  66: "Banaras Hindu University",
-  67: "Banasthali Vidyapith",
-  68: "Banda University of Agriculture and Technology",
-  69: "Bangalore University",
-  70: "Bankura University",
-  71: "Bareilly International University",
-  72: "Barkatullah University",
-  73: "Bastar Vishwavidyalaya",
-  74: "Bengaluru North University",
-  75: "Bennett University",
-  76: "Berhampur University",
-  77: "Bhagat Phool Singh Mahila Vishwavidyalaya",
-  78: "Bhagwant University",
-  79: "Bhakta Kavi Narsinh Mehta University",
-  80: "Bharath Institute of Higher Education and Research",
-  81: "Bharathiar University",
-  82: "Bharathidasan University",
-  83: "Bharati Vidyapeeth University",
-  84: "Bhartiya Skill Development University",
-  85: "Bhatkhande Music Institute",
-  86: "Bhupal Nobles University",
-  87: "Bhupendra Narayan Mandal University",
-  88: "Bidhan Chandra Krishi Vishwavidyalaya",
-  89: "Bihar Agricultural University",
-  90: "Biju Patnaik University of Technology",
-  91: "Bilaspur Vishwavidyalaya",
-  92: "Birla Institute of Technology",
-  93: "Birla Institute of Technology and Science",
-  94: "Birsa Agricultural University",
-  95: "BLDE University",
-  96: "BML Munjal University",
-  97: "Bodoland University",
-  98: "Brainware University",
-  99: "Bundelkhand University",
-  100: "C.U. Shah University",
-  101: "Calorx Teachers University",
-  102: "Career Point University",
-  103: "Central Agricultural University",
-  104: "Central Institute of Fisheries Education",
-  105: "Central University of Gujarat",
-  106: "Central University of Haryana",
-  107: "Central University of Himachal Pradesh",
-  108: "Central University of Jammu",
-  109: "Central University of Jharkhand",
-  110: "Central University of Karnataka",
-  111: "Central University of Kashmir",
-  112: "Central University of Kerala",
-  113: "Central University of Orissa",
-  114: "Central University of Punjab",
-  115: "Central University of Rajasthan",
-  116: "Central University of South Bihar",
-  117: "Central University of Tamil Nadu",
-  118: "Central University of Tibetan Studies",
-  119: "Centurion University of Technology and Management",
-  120: "CEPT University",
-  121: "Chanakya National Law University",
-  122: "Chandigarh University",
-  123: "Chandra Shekhar Azad University of Agriculture and Technology",
-  124: "Charotar University of Science and Technology",
-  125: "Chaudhary Bansi Lal University",
-  126: "Chaudhary Charan Singh Haryana Agricultural University",
-  127: "Chaudhary Charan Singh University",
-  128: "Chaudhary Devi Lal University",
-  129: "Chaudhary Ranbir Singh University",
-  130: "Chennai Mathematical Institute",
-  131: "Chhatrapati Shahu Ji Maharaj University",
-  132: "Chhattisgarh Kamdhenu Vishwavidyalaya",
-  133: "Chhattisgarh Swami Vivekananda Technical University",
-  134: "Children's University",
-  135: "Chitkara University",
-  136: "Chitkara University",
-  137: "Christ University",
-  138: "Cluster University of Jammu",
-  139: "CMJ University",
-  140: "CMR University",
-  141: "Cochin University of Science and Technology",
-  142: "Cooch Behar Panchanan Barma University",
-  143: "Cotton University",
-  144: "CSK Himachal Pradesh Krishi Vishvavidyalaya",
-  145: "Dakshina Bharat Hindi Prachar Sabha",
-  146: "Damodaram Sanjivayya National Law University",
-  147: "Datta Meghe Institute of Medical Sciences",
-  148: "DAV University",
-  149: "Davangere University",
-  150: "Dayalbagh Educational Institute",
-  151: "Dayananda Sagar University",
-  152: "Deccan College Post-Graduate and Research Institute",
-  153: "Deen Dayal Upadhyay Gorakhpur University",
-  154: "Deenbandhu Chhotu Ram University of Science and Technology",
-  155: "Delhi Pharmaceutical Sciences and Research University",
-  156: "Delhi Technological University",
-  157: "Desh Bhagat University",
-  158: "Dev Sanskriti Vishwavidyalaya",
-  159: "Devi Ahilya Vishwavidyalaya",
-  160: "Dharmsinh Desai University",
-  161: "Dhirubhai Ambani Institute of Information and Communication Technology",
-  162: "Diamond Harbour Women's University",
-  163: "Dibrugarh University",
-  164: "DIT University",
-  165: "Doon University",
-  166: "Dr K.N. Modi University",
-  167: "Dr. A.P.J Abdul Kalam University",
-  168: "Dr. A.P.J. Abdul Kalam Technical University",
-  169: "Dr. B R Ambedkar National Institute of Technology Jalandhar",
-  170: "Dr. B.R. Ambedkar University",
-  171: "Dr. B.R. Ambedkar University of Social Sciences",
-  172: "Dr. Babasaheb Ambedkar Marathwada University",
-  173: "Dr. Babasaheb Ambedkar Technological University",
-  174: "Dr. Balasaheb Sawant Konkan Krishi Vidyapeeth",
-  175: "Dr. Bhimrao Ambedkar University",
-  176: "Dr. C.V. Raman University",
-  177: "Dr. D.Y. Patil Vidyapeeth",
-  178: "Dr. Hari Singh Gour University",
-  179: "Dr. M.G.R. Educational and Research Institute",
-  180: "Dr. N.T.R. University of Health Sciences",
-  181: "Dr. Panjabrao Deshmukh Krishi Vidyapeeth",
-  182: "Dr. Rajendra Prasad Central Agriculture University",
-  183: "Dr. Ram Manohar Lohia Avadh University",
-  184: "Dr. Ram Manohar Lohiya National Law University",
-  185: "Dr. Sarvepalli Radhakrishnan Rajasthan Ayurved University",
-  186: "Dr. Shakuntala Misra Rehabilitation University",
-  187: "Dr. Vishwanath Karad MIT World Peace University",
-  188: "Dr. Y.S. Parmar University of Horticulture and Forestry",
-  189: "Dr.Y.S.R. Horticultural University",
-  190: "Dravidian University",
-  191: "EIILM University",
-  192: "Era University",
-  193: "Eternal University",
-  194: "Fakir Mohan University",
-  195: "Flame University",
-  196: "Forest Research Institute",
-  197: "G H Raisoni University",
-  198: "G.L.S. University",
-  199: "Galgotias University",
-  200: "Gandhi Institute of Technology and Management",
-  201: "Gandhigram Rural University",
-  202: "Gangadhar Meher University",
-  203: "Ganpat University",
-  204: "Garden City University",
-  205: "Gauhati University",
-  206: "Gautam Buddha University",
-  207: "GD Goenka University",
-  208: "Geetanjali University",
-  209: "GLA University",
-  210: "Glocal University",
-  211: "GNA University",
-  212: "Goa University",
-  213: "Gokhale Institute of Politics and Economics",
-  214: "Gondwana University",
-  215: "Govind Ballabh Pant University of Agriculture and Technology",
-  216: "Govind Guru Tribal University",
-  217: "Graphic Era Hill University",
-  218: "Graphic Era University",
-  219: "GSFC University",
-  220: "Gujarat Ayurved University",
-  221: "Gujarat Forensic Sciences University",
-  222: "Gujarat National Law University",
-  223: "Gujarat Technological University",
-  224: "Gujarat University",
-  225: "Gujarat University of Transplantation Sciences",
-  226: "Gujarat Vidyapith",
-  227: "Gulbarga University",
-  228: "Guru Angad Dev Veterinary and Animal Sciences University",
-  229: "Guru Ghasidas Vishwavidyalaya",
-  230: "Guru Gobind Singh Indraprastha University",
-  231: "Guru Jambheshwar University of Science and Technology",
-  232: "Guru Kashi University",
-  233: "Guru Nanak Dev University",
-  234: "Guru Ravidas Ayurved University",
-  235: "Gurukul Kangri Vishwavidyalaya",
-  236: "Harcourt Butler Technical University",
-  237: "Haridev Joshi University of Journalism and Mass Communication",
-  238: "Haryana Vishwakarma Skill University",
-  239: "Hemchandracharya North Gujarat University",
-  240: "Hemwati Nandan Bahuguna Garhwal University",
-  241: "Hemwati Nandan Bahuguna Medical University",
-  242: "Hidayatullah National Law University",
-  243: "Himachal Pradesh Technical University",
-  244: "Himachal Pradesh University",
-  245: "Himalayan Garhwal University",
-  246: "Himalayan University",
-  247: "Himgiri ZEE University",
-  248: "Hindustan Institute of Technology and Science",
-  249: "Homi Bhabha National Institute",
-  250: "ICFAI University, Dehradun",
-  251: "ICFAI University, Himachal Pradesh",
-  252: "ICFAI University, Jaipur",
-  253: "ICFAI University, Jharkhand",
-  254: "ICFAI University, Meghalaya",
-  255: "ICFAI University, Mizoram",
-  256: "ICFAI University, Nagaland",
-  257: "ICFAI University, Raipur",
-  258: "ICFAI University, Sikkim",
-  259: "ICFAI University, Tripura",
-  260: "IEC University",
-  261: "IFHE Hyderabad",
-  262: "IFTM University",
-  263: "IIHMR University",
-  264: "IIMT University",
-  265: "IMS Unison University",
-  266: "Indian Agricultural Research Institute",
-  267: "Indian Institute of Engineering Science and Technology, Shibpur",
-  268: "Indian Institute of Foreign Trade",
-  269: "Indian Institute of Information Technology Allahabad",
-  270: "Indian Institute of Information Technology and Management Gwalior",
-  271: "Indian Institute of Information Technology, Design and Manufacturing",
-  272: "Indian Institute of Information Technology, Guwahati",
-  273: "Indian Institute of Information Technology, Kalyani",
-  274: "Indian Institute of Information Technology, Kottayam",
-  275: "Indian Institute of Information Technology, Lucknow",
-  276: "Indian Institute of Information Technology, Manipur",
-  277: "Indian Institute of Information Technology, Sri City",
-  278: "Indian Institute of Information Technology, Una",
-  279: "Indian Institute of Information Technology, Vadodara",
-  280: "Indian Institute of Public Health",
-  281: "Indian Institute of Science",
-  282: "Indian Institute of Science Education and Research, Bhopal",
-  283: "Indian Institute of Science Education and Research, Kolkata",
-  284: "Indian Institute of Science Education and Research, Mohali",
-  285: "Indian Institute of Science Education and Research, Pune",
-  286: "Indian Institute of Science Education and Research, Thiruvananthapuram",
-  287: "Indian Institute of Space Science and Technology",
-  288: "Indian Institute of Teacher Education",
-  289: "Indian Institute of Technology Bhubaneswar",
-  290: "Indian Institute of Technology Bombay",
-  291: "Indian Institute of Technology Delhi",
-  292: "Indian Institute of Technology Gandhinagar",
-  293: "Indian Institute of Technology Guwahati",
-  294: "Indian Institute of Technology Hyderabad",
-  295: "Indian Institute of Technology Indore",
-  296: "Indian Institute of Technology Jodhpur",
-  297: "Indian Institute of Technology Kanpur",
-  298: "Indian Institute of Technology Kharagpur",
-  299: "Indian Institute of Technology Madras",
-  300: "Indian Institute of Technology Mandi",
-  301: "Indian Institute of Technology Patna",
-  302: "Indian Institute of Technology Roorkee",
-  303: "Indian Institute of Technology Ropar",
-  304: "Indian Institute of Technology, BHU",
-  305: "Indian Maritime University",
-  306: "Indian School of Mines",
-  307: "Indian Statistical Institute",
-  308: "Indian Veterinary Research Institute",
-  309: "Indira Gandhi Delhi Technical University for Women",
-  310: "Indira Gandhi Institute of Development Research",
-  311: "Indira Gandhi Institute of Medical Sciences, Sheikhpura",
-  312: "Indira Gandhi Krishi Vishwavidyalaya",
-  313: "Indira Gandhi National Tribal University",
-  314: "Indira Gandhi Technological and Medical Sciences University",
-  315: "Indira Gandhi University, Meerpur",
-  316: "Indira Kala Sangeet Vishwavidyalaya",
-  317: "Indraprastha Institute of Information Technology",
-  318: "Indrashil University",
-  319: "Indus International University",
-  320: "Indus University",
-  321: "Institute of Advanced Research",
-  322: "Institute of Advanced Studies in Education",
-  323: "Institute of Chemical Technology",
-  324: "Institute of Infrastructure, Technology, Research and Management",
-  325: "Integral University",
-  326: "International Institute for Population Sciences",
-  327: "International Institute of Information Technology Bangalore",
-  328: "International Institute of Information Technology, Bhubaneswar",
-  329: "International Institute of Information Technology, Hyderabad",
-  330: "International Institute of Information Technology, Naya Raipur",
-  331: "Invertis University",
-  332: "ISBM University",
-  333: "Islamic University of Science and Technology",
-  334: "ITM University Gwalior",
-  335: "ITM University Raipur",
-  336: "ITM Vocational University",
-  337: "Jadavpur University",
-  338: "Jagadguru Ramanandacharya Rajasthan Sanskrit University",
-  339: "Jagadguru Rambhadracharya Handicapped University",
-  340: "Jagadguru Sri Shivarathreeswara University",
-  341: "Jagan Nath University",
-  342: "Jagannath University",
-  343: "Jagran Lakecity University",
-  344: "Jai Narain Vyas University",
-  345: "Jai Prakash Vishwavidyalaya",
-  346: "Jain University",
-  347: "Jain Vishva Bharati Institute",
-  348: "Jaipur National University",
-  349: "Jamia Hamdard",
-  350: "Jamia Millia Islamia",
-  351: "Jananayak Chandrashekhar University",
-  352: "Janardan Rai Nagar Rajasthan Vidhyapeeth University",
-  353: "Jawaharlal Institute of Postgraduate Medical Education and Research",
-  354: "Jawaharlal Nehru Architecture and Fine Arts University",
-  355: "Jawaharlal Nehru Centre for Advanced Scientific Research",
-  356: "Jawaharlal Nehru Krishi Vishwavidyalaya",
-  357: "Jawaharlal Nehru Technological University",
-  358: "Jawaharlal Nehru Technological University, Anantapur",
-  359: "Jawaharlal Nehru Technological University, Kakinada",
-  360: "Jawaharlal Nehru University",
-  361: "Jayoti Vidyapeeth Women's University",
-  362: "Jaypee Institute of Information Technology",
-  363: "Jaypee University Anoopshahr",
-  364: "Jaypee University of Engineering and Technology",
-  365: "Jaypee University of Information Technology",
-  366: "JECRC University",
-  367: "Jharkhand Rai University",
-  368: "Jharkhand Raksha Shakti University",
-  369: "JIS University",
-  370: "Jiwaji University",
-  371: "JK Lakshmipat University",
-  372: "Jodhpur National University",
-  373: "JS University",
-  374: "JSS Science and Technology University",
-  375: "Junagadh Agricultural University",
-  376: "K L University",
-  377: "K.K. University",
-  378: "K.R. Mangalam University",
-  379: "Kadi Sarva Vishwavidyalaya",
-  380: "Kakatiya University",
-  381: "Kalasalingam University",
-  382: "Kalinga University",
-  383: "Kamdhenu University",
-  384: "Kameshwar Singh Darbhanga Sanskrit University",
-  385: "Kannada University",
-  386: "Kannur University",
-  387: "Karnatak University",
-  388: "Karnataka Janapada Vishwavidyalaya",
-  389: "Karnataka Sanskrit University",
-  390: "Karnataka State Law University",
-  391: "Karnataka State Rural Development and Panchayat Raj University",
-  392: "Karnataka State Women's University",
-  393: "Karnataka Veterinary, Animal and Fisheries Sciences University",
-  394: "Karnavati University",
-  395: "Karpagam Academy of Higher Education",
-  396: "Karunya University",
-  397: "Kavi Kulguru Kalidas Sanskrit Vishwavidyalaya",
-  398: "Kazi Nazrul University",
-  399: "Kaziranga University",
-  400: "Kerala Agricultural University",
-  401: "Kerala Kalamandalam",
-  402: "Kerala University of Fisheries and Ocean Studies",
-  403: "Kerala University of Health Sciences",
-  404: "Kerala Veterinary and Animal Sciences University",
-  405: "Khallikote University",
-  406: "Khalsa University",
-  407: "Khwaja Moinuddin Chishti Urdu, Arabi-Farsi University",
-  408: "KIIT University",
-  409: "King George's Medical University",
-  410: "KLE Technological University",
-  411: "KLE University",
-  412: "Kolhan University",
-  413: "Krantiguru Shyamji Krishna Verma Kachchh University",
-  414: "Krishna Institute of Medical Sciences",
-  415: "Krishna University",
-  416: "KSGH Music and Performing Arts University",
-  417: "Kumar Bhaskar Varma Sanskrit and Ancient Studies University",
-  418: "Kumaun University",
-  419: "Kurukshetra University",
-  420: "Kushabhau Thakre Patrakarita Avam Jansanchar University",
-  421: "Kuvempu University",
-  422: "Lakshmibai National Institute of Physical Education",
-  423: "Lakulish Yoga University",
-  424: "Lala Lajpat Rai University of Veterinary and Animal Sciences",
-  425: "Lalit Narayan Mithila University",
-  426: "Lingaya's University",
-  427: "LNCT University",
-  428: "Lovely Professional University",
-  429: "M.J.P. Rohilkhand University",
-  430: "M.S. Ramaiah University of Applied Sciences",
-  431: "Madan Mohan Malaviya University of Technology",
-  432: "Madhav University",
-  433: "Madurai Kamaraj University",
-  434: "Magadh University",
-  435: "Mahapurusha Srimanta Sankaradeva Viswavidyalaya",
-  436: "Maharaja Agrasen University",
-  437: "Maharaja Bir Bikram University",
-  438: "Maharaja Chhatrasal Bundelkhand University",
-  439: "Maharaja Ganga Singh University",
-  440: "Maharaja Krishnakumarsinhji Bhavnagar University",
-  441: "Maharaja Ranjit Singh Punjab Technical University",
-  442: "Maharaja Surajmal Brij University",
-  443: "Maharana Pratap University of Agriculture and Technology",
-  444: "Maharashtra Animal and Fishery Sciences University",
-  445: "Maharashtra National Law University, Aurangabad",
-  446: "Maharashtra National Law University, Mumbai",
-  447: "Maharashtra National Law University, Nagpur",
-  448: "Maharashtra University of Health Sciences",
-  449: "Maharishi Arvind University, Jaipur",
-  450: "Maharishi Dayanand University",
-  451: "Maharishi Mahesh Yogi Vedic Vishwavidyalaya",
-  452: "Maharishi Markandeshwar University, Mullana",
-  453: "Maharishi Markandeshwar University, Sadopur",
-  454: "Maharishi Markandeshwar University, Solan",
-  455: "Maharishi University",
-  456: "Maharishi University of Management and Technology",
-  457: "Maharshi Dayanand Saraswati University",
-  458: "Maharshi Panini Sanskrit Vishwavidyalaya",
-  459: "Mahatma Gandhi Antarrashtriya Hindi Vishwavidyalaya",
-  460: "Mahatma Gandhi Central University, Motihari",
-  461: "Mahatma Gandhi Chitrakoot Gramoday Vishwavidyalaya",
-  462: "Mahatma Gandhi Kashi Vidyapeeth",
-  463: "Mahatma Gandhi University",
-  464: "Mahatma Gandhi University of Medical Sciences and Technology",
-  465: "Mahatma Gandhi University, Meghalaya",
-  466: "Mahatma Gandhi University, Nalgonda",
-  467: "Mahatma Jyoti Rao Phoole University",
-  468: "Mahatma Phule Krishi Vidyapeeth",
-  469: "Makhanlal Chaturvedi Rashtriya Patrakarita Vishwavidyalaya",
-  470: "Malaviya National Institute of Technology, Jaipur",
-  471: "Malwanchal University",
-  472: "Manav Bharti University",
-  473: "Manav Rachna International Institute of Research and Studies",
-  474: "Manav Rachna University",
-  475: "Mandsaur University",
-  476: "Mangalayatan University",
-  477: "Mangalore University",
-  478: "Manipal University",
-  479: "Manipur Technical University",
-  480: "Manipur University",
-  481: "Manipur University of Culture",
-  482: "Manonmaniam Sundaranar University",
-  483: "Martin Luther Christian University",
-  484: "Marwadi University",
-  485: "MATS University",
-  486: "Maulana Abul Kalam Azad University of Technology",
-  487: "Maulana Azad National Institute of Technology",
-  488: "Maulana Azad National Urdu University",
-  489: "Maulana Azad University, Jodhpur",
-  490: "Maulana Mazharul Haque Arabic and Persian University",
-  491: "Medi-Caps University",
-  492: "Meenakshi Academy of Higher Education and Research",
-  493: "Mewar University",
-  494: "MGM Institute of Health Sciences",
-  495: "MIT Art Design and Technology University",
-  496: "Mizoram University",
-  497: "Mody University of Science and Technology",
-  498: "Mohammad Ali Jauhar University",
-  499: "Mohanlal Sukhadia University",
-  500: "Monad University",
-  501: "Mother Teresa Women's University",
-  502: "Motherhood University",
-  503: "Motilal Nehru National Institute of Technology",
-  504: "MVN University",
-  505: "Nagaland University",
-  506: "Nalanda University",
-  507: "NALSAR University of Law",
-  508: "Nanaji Deshmukh Veterinary Science University",
-  509: "Narendra Dev University of Agriculture and Technology",
-  510: "Narsee Monjee Institute of Management and Higher Studies",
-  511: "National Dairy Research Institute",
-  512: "National Institute of Design",
-  513: "National Institute of Fashion Technology",
-  514: "National Institute of Food Technology Entrepreneurship and Management",
-  515: "National Institute of Mental Health and Neuro Sciences",
-  516: "National Institute of Pharmaceutical Education and Research, Ahmedabad",
-  517: "National Institute of Pharmaceutical Education and Research, Guwahati",
-  518: "National Institute of Pharmaceutical Education and Research, Hajipur",
-  519: "National Institute of Pharmaceutical Education and Research, Hyderabad",
-  520: "National Institute of Pharmaceutical Education and Research, Kolkata",
-  521: "National Institute of Pharmaceutical Education and Research, Mohali",
-  522: "National Institute of Pharmaceutical Education and Research, Rae Bareli",
-  523: "National Institute of Technology, Agartala",
-  524: "National Institute of Technology, Arunachal Pradesh",
-  525: "National Institute of Technology, Calicut",
-  526: "National Institute of Technology, Delhi",
-  527: "National Institute of Technology, Durgapur",
-  528: "National Institute of Technology, Goa",
-  529: "National Institute of Technology, Hamirpur",
-  530: "National Institute of Technology, Jamshedpur",
-  531: "National Institute of Technology, Karnataka",
-  532: "National Institute of Technology, Kurukshetra",
-  533: "National Institute of Technology, Manipur",
-  534: "National Institute of Technology, Meghalaya",
-  535: "National Institute of Technology, Mizoram",
-  536: "National Institute of Technology, Nagaland",
-  537: "National Institute of Technology, Patna",
-  538: "National Institute of Technology, Puducherry",
-  539: "National Institute of Technology, Raipur",
-  540: "National Institute of Technology, Rourkela",
-  541: "National Institute of Technology, Sikkim",
-  542: "National Institute of Technology, Silchar",
-  543: "National Institute of Technology, Srinagar",
-  544: "National Institute of Technology, Tiruchirappalli",
-  545: "National Institute of Technology, Uttarakhand",
-  546: "National Institute of Technology, Warangal",
-  547: "National Law Institute University",
-  548: "National Law School of India University",
-  549: "National Law University",
-  550: "National Law University and Judicial Academy",
-  551: "National Law University, Jodhpur",
-  552: "National Law University, Orissa",
-  553: "National University of Educational Planning and Administration",
-  554: "National University of Study and Research in Law",
-  555: "Nava Nalanda Mahavihara",
-  556: "Navrachana University",
-  557: "Navsari Agricultural University",
-  558: "Nehru Gram Bharati Vishwavidyalaya",
-  559: "NIILM University",
-  560: "NIIT University",
-  561: "Nilamber-Pitamber University",
-  562: "NIMS University",
-  563: "Nirma University of Science and Technology",
-  564: "NITTE University",
-  565: "Nizam's Institute of Medical Sciences",
-  566: "Noida International University",
-  567: "North East Frontier Technical University",
-  568: "North Eastern Hill University",
-  569: "North Eastern Regional Institute of Science and Technology",
-  570: "North Maharashtra University",
-  571: "North Orissa University",
-  572: "O.P. Jindal Global University",
-  573: "O.P. Jindal University",
-  574: "OPJS University",
-  575: "Oriental University",
-  576: "Orissa University of Agriculture and Technology",
-  577: "Osmania University",
-  578: "P P Savani University",
-  579: "P.K. University",
-  580: "Pacific Medical University",
-  581: "Pacific University, India",
-  582: "Padmashree Dr. D.Y. Patil Vidyapith",
-  583: "Palamuru University",
-  584: "Pandit Deendayal Petroleum University",
-  585: "Pandit Deendayal Upadhyaya Shekhawati University",
-  586: "Pandit Ravishankar Shukla University",
-  587: "Panjab University",
-  588: "Parul University",
-  589: "Patna University",
-  590: "PDM University",
-  591: "PDPM Indian Institute of Information Technology, Design and Manufacturing",
-  592: "PEC University of Technology",
-  593: "People's University",
-  594: "Periyar Maniammai University",
-  595: "Periyar University",
-  596: "PES University",
-  597: "Plastindia International University",
-  598: "Pondicherry University",
-  599: "Ponnaiyan Ramajayam Institute of Science and Technology",
-  600: "Poornima University",
-  601: "Post Graduate Institute of Medical Education and Research",
-  602: "Potti Sreeramulu Telugu University",
-  603: "Pragyan International University",
-  604: "Pratap University",
-  605: "Pravara Institute of Medical Sciences",
-  606: "Presidency University",
-  607: "Presidency University",
-  608: "Professor Jayashankar Telangana State Agricultural University",
-  609: "Pt. Bhagwat Dayal Sharma University of Health Sciences",
-  610: "Punjab Agricultural University",
-  611: "Punjab Technical University",
-  612: "Punjabi University",
-  613: "Quantum University",
-  614: "Rabindra Bharati University",
-  615: "Raffles University",
-  616: "Rai Technology University",
-  617: "Rai University",
-  618: "Raiganj University",
-  619: "Raj Rishi Bharthari Matsya University",
-  620: "Raja Mansingh Tomar Music and Arts University",
-  621: "Rajasthan Technical University Kota",
-  622: "Rajasthan University of Health Sciences",
-  623: "Rajasthan University of Veterinary and Animal Sciences",
-  624: "Rajiv Gandhi Institute of Petroleum Technology",
-  625: "Rajiv Gandhi National Aviation University",
-  626: "Rajiv Gandhi National Institute of Youth Development",
-  627: "Rajiv Gandhi National University of Law",
-  628: "Rajiv Gandhi Proudyogiki Vishwavidyalaya",
-  629: "Rajiv Gandhi University",
-  630: "Rajiv Gandhi University of Health Sciences",
-  631: "Rajiv Gandhi University of Knowledge Technologies",
-  632: "Rajmata Vijayaraje Scindia Krishi Vishwavidyalaya",
-  633: "Raksha Shakti University",
-  634: "Rama Devi Women's University",
-  635: "Rama University",
-  636: "Ramakrishna Mission Vivekananda University",
-  637: "Ranchi University",
-  638: "Rani Channamma University, Belagavi",
-  639: "Rani Durgavati Vishwavidyalaya",
-  640: "Rani Lakshmi Bai Central Agricultural University",
-  641: "Ras Bihari Bose Subharti University",
-  642: "Rashtrasant Tukadoji Maharaj Nagpur University",
-  643: "Rashtriya Sanskrit Sansthan University",
-  644: "Rashtriya Sanskrit Vidyapeetha",
-  645: "Ravenshaw University",
-  646: "Rayalaseema University",
-  647: "Rayat-Bahra University",
-  648: "REVA University",
-  649: "RIMT University",
-  650: "RK University",
-  651: "RKDF University",
-  652: "Sai Nath University",
-  653: "Sai Tirupati University",
-  654: "Sam Higginbottom Institute of Agriculture, Technology and Sciences",
-  655: "Sambalpur University",
-  656: "Sampurnanand Sanskrit Vishvavidyalaya",
-  657: "Sanchi University of Buddhist-Indic Studies",
-  658: "Sandip University",
-  659: "Sandip University, Sijoul",
-  660: "Sangai International University",
-  661: "Sangam University",
-  662: "Sanjay Gandhi Post Graduate Institute of Medical Sciences",
-  663: "Sanjay Ghodawat University",
-  664: "Sankalchand Patel University",
-  665: "Sanskriti University",
-  666: "Sant Baba Bhag Singh University",
-  667: "Sant Gadge Baba Amravati University",
-  668: "Sant Longowal Institute of Engineering and Technology",
-  669: "Santosh University",
-  670: "Sardar Patel University",
-  671: "Sardar Patel University of Police, Security and Criminal Justice",
-  672: "Sardar Vallabhbhai National Institute of Technology, Surat",
-  673: "Sardar Vallabhbhai Patel University of Agriculture and Technology",
-  674: "Sardarkrushinagar Dantiwada Agricultural University",
-  675: "Sarguja University",
-  676: "Sarla Birla University",
-  677: "Sarvepalli Radhakrishnan University",
-  678: "SASTRA University",
-  679: "Satavahana University",
-  680: "Sathyabama University",
-  681: "Saurashtra University",
-  682: "Saveetha University",
-  683: "Savitribai Phule Pune University",
-  684: "School of Planning and Architecture, Bhopal",
-  685: "School of Planning and Architecture, Delhi",
-  686: "School of Planning and Architecture, Vijayawada",
-  687: "Seacom Skills University",
-  688: "Sharda University",
-  689: "Sharnbasva University",
-  690: "Sher-e-Kashmir University of Agricultural Sciences and Technology of Jammu",
-  691: "Sher-e-Kashmir University of Agricultural Sciences and Technology of Kashmir",
-  692: "Sher-i-Kashmir Institute of Medical Sciences",
-  693: "Shiv Nadar University",
-  694: "Shivaji University",
-  695: "Shobhit University",
-  696: "Shoolini University of Biotechnology and Management Sciences",
-  697: "Shree Guru Gobind Singh Tricentenary University",
-  698: "Shree Somnath Sanskrit University",
-  699: "Shreemati Nathibai Damodar Thackersey Women's University",
-  700: "Shri Govind Guru University",
-  701: "Shri Guru Ram Rai Education Mission",
-  702: "Shri Guru Ram Rai University",
-  703: "Shri Jagannath Sanskrit Vishvavidyalaya",
-  704: "Shri Jagdishprasad Jhabrmal Tibrewala University",
-  705: "Shri Lal Bahadur Shastri Rashtriya Sanskrit Vidyapeetha",
-  706: "Shri Mata Vaishno Devi University",
-  707: "Shri Ramswaroop Memorial University",
-  708: "Shri Vaishnav Vidyapeeth Vishwavidyalaya",
-  709: "Shri Venkateshwara University",
-  710: "Shridhar University",
-  711: "Siddharth University",
-  712: "Sidho Kanho Birsha University",
-  713: "Sido Kanhu Murmu University",
-  714: "Sikkim Manipal University",
-  715: "Sikkim University",
-  716: "Siksha O Anusandhan University",
-  717: "Singhania University",
-  718: "Sir Padampat Singhania University",
-  719: "South Asian University",
-  720: "Spicer Adventist University",
-  721: "Sree Chitra Thirunal Institute of Medical Sciences and Technology",
-  722: "Sree Sankaracharya University of Sanskrit",
-  723: "Sri Balaji Vidyapeeth",
-  724: "Sri Chandrasekharendra Saraswathi Viswa Mahavidyalaya",
-  725: "Sri Dev Suman Uttarakhand University",
-  726: "Sri Devaraj Urs Academy of Higher Education and Research",
-  727: "Sri Guru Granth Sahib World University",
-  728: "Sri Guru Ram Das University of Health Sciences",
-  729: "Sri Konda Laxman Telangana State Horticultural University",
-  730: "Sri Krishnadevaraya University",
-  731: "Sri Padmavati Mahila Visvavidyalayam",
-  732: "Sri Ramachandra University",
-  733: "Sri Sai University",
-  734: "Sri Sathya Sai Institute of Higher Learning",
-  735: "Sri Satya Sai University of Technology and Medical Sciences",
-  736: "Sri Siddhartha Academy of Higher Education",
-  737: "Sri Sri University",
-  738: "Sri Venkateswara Institute of Medical Sciences",
-  739: "Sri Venkateswara University",
-  740: "Sri Venkateswara Vedic University",
-  741: "Sri Venkateswara Veterinary University",
-  742: "Srimanta Sankaradeva University of Health Sciences",
-  743: "Srinivas University",
-  744: "SRM Institute of Science and Technology",
-  745: "SRM University Haryana",
-  746: "SRM University, Sikkim",
-  747: "St. Joseph University",
-  748: "St. Peter's University",
-  749: "St. Xavier's University",
-  750: "Starex University",
-  751: "State University of Performing and Visual Arts",
-  752: "Suamandeep Vidyapeeth",
-  753: "Sunrise University",
-  754: "Suresh Gyan Vihar University",
-  755: "Swami Keshwanand Rajasthan Agricultural University",
-  756: "Swami Rama Himalayan University",
-  757: "Swami Ramanand Teerth Marathwada University",
-  758: "Swami Vivekanand Subharti University",
-  759: "Swami Vivekanand University",
-  760: "Swami Vivekananda Yoga Anusandhana Samsthana",
-  761: "Swarnim Gujarat Sports University",
-  762: "Swarnim Startup and Innovation University",
-  763: "Symbiosis International University",
-  764: "Symbiosis University of Applied Sciences",
-  765: "Tamil Nadu Agricultural University",
-  766: "Tamil Nadu Dr Ambedkar Law University",
-  767: "Tamil Nadu Dr. M.G.R.Medical University",
-  768: "Tamil Nadu Fisheries University",
-  769: "Tamil Nadu Music and Fine Arts University",
-  770: "Tamil Nadu National Law School",
-  771: "Tamil Nadu Physical Education and Sports University",
-  772: "Tamil Nadu Teacher Education University",
-  773: "Tamil Nadu Veterinary and Animal Sciences University",
-  774: "Tamil University",
-  775: "Tantia University",
-  776: "Tata Institute of Fundamental Research",
-  777: "Tata Institute of Social Sciences",
-  778: "TeamLease Skills University",
-  779: "Techno India University",
-  780: "Teerthanker Mahaveer University",
-  781: "Telangana University",
-  782: "TERI University",
-  783: "Tezpur University",
-  784: "Thapar Institute of Engineering and Technology",
-  785: "The English and Foreign Languages University",
-  786: "The IIS University",
-  787: "The Indian Law Institute",
-  788: "The LNM Institute of Information Technology",
-  789: "The Maharaja Sayajirao University of Baroda",
-  790: "The National University of Advanced Legal Studies",
-  791: "The Neotia University",
-  792: "The Northcap University",
-  793: "The Sanskrit College and University",
-  794: "The West Bengal National University of Juridical Sciences",
-  795: "Thiruvalluvar University",
-  796: "Thunchath Ezhuthachan Malayalam University",
-  797: "Tilak Maharashtra Vidyapeeth",
-  798: "Tilka Manjhi Bhagalpur University",
-  799: "TransDisciplinary University",
-  800: "Tripura University",
-  801: "Tumkur University",
-  802: "U.P. Pandit Deen Dayal Upadhyaya Pashu Chikitsa Vigyan Vishwavidyalaya Evam Go-Anusandhan Sansthan",
-  803: "UKA Tarsadia University",
-  804: "University of Agricultural and Horticultural Sciences, Shivamogga",
-  805: "University of Agricultural Sciences, Bangalore",
-  806: "University of Agricultural Sciences, Dharwad",
-  807: "University of Agricultural Sciences, Raichur",
-  808: "University of Allahabad",
-  809: "University of Burdwan",
-  810: "University of Calcutta",
-  811: "University of Calicut",
-  812: "University of Delhi",
-  813: "University of Engineering and Management, Kolkata",
-  814: "University of Gour Banga",
-  815: "University of Horticultural Sciences, Bagalkot",
-  816: "University of Hyderabad",
-  817: "University of Jammu",
-  818: "University of Kalyani",
-  819: "University of Kashmir",
-  820: "University of Kerala",
-  821: "University of Kota",
-  822: "University of Lucknow",
-  823: "University of Madras",
-  824: "University of Mumbai",
-  825: "University of Mysore",
-  826: "University of North Bengal",
-  827: "University of Patanjali",
-  828: "University of Petroleum and Energy Studies",
-  829: "University of Rajasthan",
-  830: "University of Science and Technology, Meghalaya",
-  831: "University of Solapur",
-  832: "University of Technology",
-  833: "University of Technology and Management",
-  834: "Usha Martin University",
-  835: "Utkal University",
-  836: "Utkal University of Culture",
-  837: "Uttar Banga Krishi Viswavidyalaya",
-  838: "Uttar Pradesh University of Medical Sciences",
-  839: "Uttarakhand Aawasiya Vishwavidyalaya",
-  840: "Uttarakhand Ayurved University",
-  841: "Uttarakhand Technical University",
-  842: "Uttarakhand University of Horticulture and Forestry",
-  843: "Uttaranchal Sanskrit University",
-  844: "Uttaranchal University",
-  845: "Vasantrao Naik Marathwada Krishi Vidyapeeth",
-  846: "Veer Bahadur Singh Purvanchal University",
-  847: "Veer Kunwar Singh University",
-  848: "Veer Narmad South Gujarat University",
-  849: "Veer Surendra Sai University of Technology",
-  850: "Vel Tech Dr.RR and Dr.SR Technical University",
-  851: "Vels University",
-  852: "Vidyasagar University",
-  853: "Vignan University",
-  854: "Vijayanagara Sri Krishnadevaraya University",
-  855: "Vikram University",
-  856: "Vikrama Simhapuri University",
-  857: "Vinayaka Missions Sikkim University",
-  858: "Vinayaka Missions University",
-  859: "Vinoba Bhave University",
-  860: "Vishwakarma University",
-  861: "Visva-Bharati University",
-  862: "Visvesvaraya National Institute of Technology",
-  863: "Visvesvaraya Technological University",
-  864: "VIT Bhopal University",
-  865: "VIT University",
-  866: "Vivekananda Global University",
-  867: "West Bengal State University",
-  868: "West Bengal University of Animal and Fishery Sciences",
-  869: "West Bengal University of Health Sciences",
-  870: "West Bengal University of Teachers' Training, Education Planning and Administration",
-  871: "William Carey University, India",
-  872: "William Carey University, Meghalaya",
-  873: "Xavier University Bhubaneswar",
-  874: "YBN University",
-  875: "Yenepoya University",
-  876: "YMCA University of Science and Technology",
-  877: "Yogi Vemana University",
-  878: "Indian Institute of Technology Dharwad",
-  879: "Indian Institute of Technology Jammu",
-  880: "Dayananda Sagar College of Engineering",
-  881: "Army Institute of Technology, Pune",
-  882: "Mahindra École Centrale, Hyderabad",
-  883: "JSS Academy of Technical Education, Noida",
-  884: "Shri Govindram Seksaria Institute of Technology and Science, Indore",
-  885: "Indian Institute of Information Technology, Nagpur",
-  886: "Institute of Engineering and Technology, DAVV, Indore",
-  887: "Gyan Ganga Institute of Technology and Sciences, Jabalpur",
-  888: "Galgotias University, Greater Noida",
-  889: "Oxford University",
-  890: "Yeshwantrao Chavan College of Engineering, Nagpur",
-  891: "Sir M. Visvesvaraya Institute of Technology",
-  892: "D. J. Sanghvi College Of Engineering",
-  893: "G.Narayanamma Institute of Technology and Science",
-  894: "Bhagwan Parshuram Institute of Technology",
-  895: "Indian Institute of Information Technology, Surat",
-  896: "G. H. Patel College of Engineering & Technology",
-  897: "G. H. Raisoni College of Engineering, Nagpur",
-  898: "Arya College of Engineering and I.T.",
-  899: "DPG Institute of Technology and Management",
-  900: "DAV Institute of Engineering and Technology",
-  901: "Netaji Subhas University of Technology, Delhi",
-  902: "Ambedkar Institute of Advanced Communication Technologies and Research",
-  903: "Maharaja Surajmal Institute Of Technology, New Delhi",
-  904: "Institute of Engineering and Technology, Lucknow",
-  905: "Mallareddy college of Engineering and Technology, Hyderabad",
-  906: "Aravali College of Engineering and Management",
-  907: "Hitkarini College Of Engineering And Technology, Jabalpur",
-  908: "RV College of Engineering, Bengaluru",
-};
-
-var collegesArr = [];
-
-for (var i in colleges) {
-  collegesArr.push(colleges[i]);
-}
-collegesArr = collegesArr.sort();
-export default collegesArr;
