--- src/components/dashboard/StudentDashboard.scss
@@ -1,359 +1,0 @@
-////////////  WARNING \\\\\\\\\\\\\\\\\\\\\\
-// CURRENTLY NOT BEING USED BEING USED BY ANY OF THE COMPONENTS IN PRODUCTION
-// FOR DASHBOARD SYTLES - REFER dashboard.scss
-
-// import using ~
-// Import only what you need from Bulma
-@import "~bulma/sass/utilities/_all.sass";
-@import "~bulma/sass/base/generic.sass";
-@import "~bulma/sass/base/helpers.sass";
-@import "~bulma/sass/grid/columns.sass";
-@import "~bulma/sass/elements/button.sass";
-@import "~bulma/sass/elements/container.sass";
-@import "~bulma/sass/elements/image.sass";
-@import "~bulma/sass/form/_all.sass";
-@import "~bulma/sass/components/navbar.sass";
-@import "~bulma/sass/layout/hero.sass";
-@import "~bulma/sass/layout/section.sass";
-
-// FONT SIZES
-
-$head-1: 3.052rem;
-$head-2: 2.441rem;
-$head-3: 1.953rem;
-$head-4: 1.563rem;
-$head-5: 1.25rem;
-
-$border: 12px;
-$border-small: 6px;
-
-// VARIABLES
-$dashboard-font: Route159, Vegur;
-$background-color-page: #26292b;
-$header-color: #b57aff;
-
-$project-stats-header: #8794a1;
-$project-card-header: #c1d3e5;
-$project-buttons: #3e4345, 100%;
-
-$text-color: white;
-// ---------
-
-// FONTS
-
-@font-face {
-  font-family: "Route159";
-  src: local("Route159"),
-    url(../../../fonts/Route159-Regular.otf) format("opentype");
-}
-@font-face {
-  font-family: "Vegur";
-
-  src: local("Vegur"), url(../../../fonts/Vegur-Regular.otf) format("opentype");
-}
-// -----
-
-* {
-  text-rendering: optimizeLegibility;
-  box-sizing: border-box;
-}
-
-html {
-  margin: 0;
-}
-
-body {
-  margin: 0;
-  font-family: Vegur, Route159;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}
-
-code {
-  font-family: source-code-pro, Menlo, Monaco, Consolas, "Courier New",
-    monospace;
-}
-
-button {
-  background: #fff;
-  border: none;
-}
-
-.App {
-  .hero {
-    h1 {
-      font-size: 5rem;
-    }
-  }
-}
-
-.dashboard {
-  background: $background-color-page;
-  // background-size: fill;
-  // background-repeat: no-repeat;
-
-  color: #5f6e7d;
-
-  .intro {
-    padding: 10rem 0 0;
-    display: flex;
-    justify-content: space-between;
-    margin: 0px;
-    margin-left: 5rem;
-    h1 {
-      font-size: 5rem;
-    }
-    h2 {
-      color: #5f6e7d;
-    }
-
-    .data-panel {
-      h1 {
-        margin: 0 0 1rem 0;
-      }
-
-      .data-cards {
-        display: flex;
-        .data-card {
-          padding: 12px 30px 26px;
-          margin-right: 5rem;
-          box-shadow: -25px -25px 50px #fdfeff, 25px 25px 50px #dee2eb;
-          border-radius: $border;
-          text-align: center;
-          h1,
-          h2 {
-            margin: 0;
-          }
-          h1 {
-            margin-bottom: -1rem;
-          }
-        }
-      }
-    }
-
-    .profile-panel {
-      align-self: center;
-      font-size: 1.4rem;
-      margin-top: 1rem;
-      width: 30%;
-      text-align: center;
-      padding: 1rem;
-
-      box-shadow: inset 10px 23px 63px -4px rgba(219, 226, 235, 1);
-      // border-radius: 20px;
-      border-bottom-left-radius: 20px;
-      border-top-left-radius: 20px;
-      img {
-        border-radius: 6px;
-        margin-bottom: 1rem;
-        padding: 1rem;
-        max-width: 250px;
-        max-height: 250px;
-      }
-      p {
-        margin: 0rem;
-      }
-    }
-  }
-
-  section {
-    margin: 5rem auto;
-  }
-
-  .commits {
-    table {
-      border-collapse: collapse;
-      font-size: 20px;
-      width: 100%;
-      th {
-        color: #5f6e7d;
-        border: none;
-      }
-      td {
-        border: none;
-        color: #5f6e7d;
-      }
-      td:nth-child(3) {
-        color: #5f6e7d;
-      }
-      tr:hover {
-        background: rgba(#a0a3f4, 0.2);
-        td {
-          border: none;
-          color: black;
-        }
-      }
-    }
-  }
-
-  .projects,
-  .share-links {
-    .links {
-      font-size: 20px;
-      a {
-        color: #5f6e7d;
-        padding: 0;
-        margin-right: 2rem;
-        &:hover {
-          color: black;
-        }
-      }
-    }
-
-    .project-card-list {
-      display: flex;
-      flex-direction: row;
-      justify-content: space-around;
-      flex-wrap: wrap;
-
-      .project-card {
-        display: flex;
-        flex-direction: column;
-        flex-basis: 0.33;
-
-        margin: 1rem;
-        margin-right: auto;
-
-        .header-div {
-          display: flex;
-          flex-direction: row;
-
-          align-items: center;
-
-          .project-card-avatar {
-            height: 40px;
-            width: 40px;
-            margin-top: 1rem;
-            margin-right: 2rem;
-            margin-left: 1rem;
-            // padding-left: 1rem;
-
-            border-radius: 50%;
-          }
-
-          p {
-            color: black;
-            font-size: 36px;
-            text-align: center;
-            margin: 0px;
-            padding-top: 1rem;
-          }
-        }
-
-        box-shadow: -25px -25px 50px #fdfeff, 25px 25px 50px #dee2eb;
-        border-radius: $border;
-
-        width: 350px;
-        height: 150px;
-
-        .project-buttons {
-          display: flex;
-          justify-content: space-evenly;
-        }
-      }
-
-      .add-project-card-small {
-        max-width: 350px;
-        max-height: 200px;
-
-        min-width: 350px;
-        min-height: 120px;
-        display: flex;
-        flex-direction: row;
-        justify-content: center;
-        align-items: center;
-
-        a {
-          display: flex;
-          align-items: center;
-          color: #8794a1;
-          h4 {
-            text-align: center;
-            font-size: 32px;
-          }
-
-          text {
-            padding-left: 1%;
-            font-family: Inter;
-            font-style: normal;
-            font-weight: 600;
-            font-size: 120px;
-
-            color: #8794a1;
-            margin: 5px;
-            padding-bottom: 5px;
-            align-items: center;
-          }
-        }
-
-        a:hover {
-          color: #363636;
-
-          text {
-            color: #363636;
-          }
-        }
-
-        a:active {
-          color: black;
-
-          text {
-            color: black;
-          }
-        }
-      }
-    }
-  }
-
-  .resource-card {
-    padding: 2% 5% 2% 5%;
-
-    .resource-header {
-      b {
-        font-size: 32px;
-        font-family: Vegur;
-        color: #5f6e7d;
-      }
-    }
-
-    .avatar-resource {
-      height: 40px;
-      width: 40px;
-      border-radius: 50%;
-      margin: 5%;
-      &:hover {
-        opacity: 0.5;
-      }
-    }
-
-    .table {
-      border-collapse: collapse;
-      font-size: 20px;
-      width: 100%;
-      th {
-        color: #5f6e7d;
-        border: none;
-        padding-top: 2%;
-      }
-      td {
-        border: none;
-        color: #5f6e7d;
-        align-items: center;
-
-        p {
-          margin: 2% 0 2% 0;
-        }
-      }
-      td:nth-child(3) {
-        color: #5f6e7d;
-      }
-      tr:hover {
-        background: rgba(#a0a3f4, 0.2);
-        td {
-          border: none;
-          color: black;
-        }
-      }
-    }
-  }
-}
