+++ src/components/indivStats/studentStats/StudentStats.js
@@ -33,7 +33,7 @@ function fetch_calls(link) {
       return err;
     });
 }
-export default function NewStudentDashboard() {
+export default function StudentDashboard() {
   const [fullName, setFullName] = useState("");
   const [collegeName, setCollegeName] = useState("");
   const [username, setUsername] = useState("");
